+++ packages/editor/src/components/visual-editor/index.js
@@ -41,6 +41,7 @@ import {
 import { useZoomOutModeExit } from './use-zoom-out-mode-exit';
 import { usePaddingAppender } from './use-padding-appender';
 import { useEditContentOnlySectionExit } from './use-edit-content-only-section-exit';
+import { SyncConnectionModal } from '../sync-connection-modal';
 
 const {
 	LayoutStyle,
@@ -409,6 +410,7 @@ function VisualEditor( {
 				}
 			) }
 		>
+			<SyncConnectionModal />
 			<ResizableEditor enableResizing={ enableResizing } height="100%">
 				<BlockCanvas
 					shouldIframe={ ! disableIframe }