+++ packages/commands/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+-   Add `category` property to command config, used to visually differentiate commands in the Command Palette.
+
 ## 1.39.0 (2026-01-29)
 
 ## 1.38.0 (2026-01-16)